synapse grid
